[api] Fetch Heartrate data from Garmin
